Clear Interval if the component doesn't have the $el property
Using the vuelidate plugin creates components that do not have an $el property, but still have _hasMetaInfo set to true, which results in the setInterval never being cleared. There may be other plugins that create similar components.

Changing the line to if (this.$el && this.$el.offsetParent !== null) return would resolve this issue.