Update README.md
Update to explain what vue-meta is not. I have had 2 different software companies propose this as a solution to implementing dynamic metadata for Facebook/Twitter. I would suggest even making this disclaimer more emphatic. As it is, here is a suggested addition. Please consider it as a token to avoiding wasted effort and frustration by others facing this problem.